server/requirements.txt
fastapi==0.135.1
uvicorn==0.29.0
python-dotenv==1.0.0
python-multipart==0.0.22
pytest==8.1.2
httpx==0.28.1
